Current state; Improvements in all areas of the dualis implementation code

+++ lib/dualis/model/study_grades.dart
@@ -7,8 +7,8 @@ class StudyGrades {
 
   final double gpaTotal;
   final double gpaMainModules;
-  final double creditsTotal;
-  final double creditsGained;
+  final int creditsTotal;
+  final int creditsGained;
 
   StudyGrades(
     this.semesters,
